refactor(core): remove barrel files
