+++ packages/core/src/components/ConnectionLine/index.ts
@@ -2,8 +2,9 @@ import type { DefineComponent } from 'vue'
 import { defineComponent, h, inject } from 'vue'
 import type { ConnectionLineProps } from '../../types'
 import { ConnectionLineType, ConnectionMode, Position } from '../../types'
-import { getHandlePosition, getMarkerId } from '../../utils'
-import { useVueFlow } from '../../composables'
+import { getHandlePosition } from '../../utils/edge'
+import { getMarkerId } from '../../utils/graph'
+import { useVueFlow } from '../../composables/useVueFlow'
 import { Slots } from '../../context'
 import { getBezierPath, getSimpleBezierPath, getSmoothStepPath } from '../Edges/utils'
