--- packages/core/src/components/Edges/index.ts
@@ -1,10 +1,0 @@
-export { default as BaseEdge } from './BaseEdge.vue'
-export { default as BezierEdge } from './BezierEdge'
-export { default as SimpleBezierEdge } from './SimpleBezierEdge'
-export { default as StepEdge } from './StepEdge'
-export { default as SmoothStepEdge } from './SmoothStepEdge'
-export { default as StraightEdge } from './StraightEdge'
-export { default as EdgeAnchor } from './EdgeAnchor'
-export { default as EdgeText } from './EdgeText.vue'
-export { default as EdgeWrapper } from './EdgeWrapper'
-export { default as EdgeLabelRenderer } from './EdgeLabelRenderer.vue'
