--- packages/core/src/components/Nodes/index.ts
@@ -1,4 +1,0 @@
-export { default as DefaultNode } from './DefaultNode'
-export { default as InputNode } from './InputNode'
-export { default as OutputNode } from './OutputNode'
-export { default as NodeWrapper } from './NodeWrapper'
