--- packages/core/src/components/index.ts
@@ -1,6 +1,0 @@
-export * from './Nodes'
-export * from './Edges'
-export { default as ConnectionLine } from './ConnectionLine'
-export { default as Handle } from './Handle/Handle.vue'
-export { default as NodesSelection } from './NodesSelection/NodesSelection.vue'
-export { default as UserSelection } from './UserSelection/UserSelection.vue'
