+++ packages/core/src/container/NodeRenderer/NodeRenderer.vue
@@ -1,11 +1,11 @@
 <script lang="ts" setup>
 import { getCurrentInstance, inject, nextTick, onBeforeUnmount, onMounted, ref, resolveComponent } from 'vue'
 import { until } from '@vueuse/core'
-import { NodeWrapper } from '../../components'
+import NodeWrapper from '../../components/Nodes/NodeWrapper'
 import type { GraphNode, HandleConnectable, NodeComponent } from '../../types'
 import { Slots } from '../../context'
-import { useVueFlow } from '../../composables'
-import { ErrorCode, VueFlowError } from '../../utils'
+import { useVueFlow } from '../../composables/useVueFlow'
+import { ErrorCode, VueFlowError } from '../../utils/errors'
 
 const slots = inject(Slots)
