--- packages/core/src/container/index.ts
@@ -1,6 +1,0 @@
-export { default as VueFlow } from './VueFlow/VueFlow.vue'
-export { default as EdgeRenderer } from './EdgeRenderer/EdgeRenderer.vue'
-export { default as Marker } from './EdgeRenderer/Marker.vue'
-export { default as MarkerDefinitions } from './EdgeRenderer/MarkerDefinitions.vue'
-export { default as NodeRenderer } from './NodeRenderer/NodeRenderer.vue'
-export { default as Viewport } from './Viewport/Viewport.vue'
