+++ packages/core/src/index.ts
@@ -9,16 +9,21 @@ export { default as Handle } from './components/Handle/Handle.vue'
 
 export { default as Panel } from './components/Panel/Panel.vue'
 
-export {
-  StraightEdge,
-  StepEdge,
-  BezierEdge,
-  SimpleBezierEdge,
-  SmoothStepEdge,
-  BaseEdge,
-  EdgeText,
-  EdgeLabelRenderer,
-} from './components/Edges'
+export { default as StraightEdge } from './components/Edges/StraightEdge'
+
+export { default as StepEdge } from './components/Edges/StepEdge'
+
+export { default as BezierEdge } from './components/Edges/BezierEdge'
+
+export { default as SimpleBezierEdge } from './components/Edges/SimpleBezierEdge'
+
+export { default as SmoothStepEdge } from './components/Edges/SmoothStepEdge'
+
+export { default as BaseEdge } from './components/Edges/BaseEdge.vue'
+
+export { default as EdgeText } from './components/Edges/EdgeText.vue'
+
+export { default as EdgeLabelRenderer } from './components/Edges/EdgeLabelRenderer.vue'
 
 export {
   getBezierPath,
