--- packages/core/src/store/index.ts
@@ -1,4 +1,0 @@
-export * from './actions'
-export * from './getters'
-export * from './hooks'
-export * from './state'
